requirements.txt
numpy==1.26.0
Pillow==10.0.1
tqdm==4.66.1
